+++ GetARock/GetARock/Screen/AddGathering/AddGatheringViewController.swift
@@ -48,8 +48,7 @@ class AddGatheringViewController: UIViewController {
     }
 
     deinit {
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
+        removeObserversForKeyboardShow()
     }
 
     // MARK: - Method
@@ -107,7 +106,7 @@ class AddGatheringViewController: UIViewController {
         hostBandNameLabel.text = hostBandName
         dateTimePicker.minimumDate = Date()
         titleTextField.becomeFirstResponder()
-        getKeyboardNotification()
+        addObeserversForKeyboardShow()
     }
 
     private func setDelegate() {
@@ -150,52 +149,6 @@ extension AddGatheringViewController: LocationSearchViewControllerDelegate {
             locationLabel.text = gatheringAddress
             locationLabel.textColor = .white
         }
-    }
-}
-
-// MARK: - KeyboardControl
-
-extension AddGatheringViewController {
-    private func getKeyboardNotification() {
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(keyboardWillShow(_:)),
-            name: UIResponder.keyboardWillShowNotification,
-            object: nil
-        )
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(keyboardWillHide(_:)),
-            name: UIResponder.keyboardWillHideNotification,
-            object: nil
-        )
-    }
-
-    @objc func keyboardWillShow(_ sender: Notification) {
-        guard let userInfo: NSDictionary = sender.userInfo as NSDictionary?,
-              let keyboardFrame: NSValue = userInfo.value(forKey: UIResponder.keyboardFrameEndUserInfoKey) as? NSValue else {
-                  return
-              }
-        let keyboardRectangle = keyboardFrame.cgRectValue
-        let keyboardHeight = keyboardRectangle.height
-
-        let contentInset = UIEdgeInsets(
-            top: 0.0,
-            left: 0.0,
-            bottom: keyboardHeight,
-            right: 0.0)
-        scrollView.contentInset = contentInset
-        scrollView.scrollIndicatorInsets = contentInset
-    }
-
-    @objc func keyboardWillHide(_ sender: Notification) {
-        let contentInset = UIEdgeInsets(
-                top: 0.0,
-                left: 0.0,
-                bottom: 0.0,
-                right: 0.0)
-            scrollView.contentInset = contentInset
-            scrollView.scrollIndicatorInsets = contentInset
     }
 }
