+++ GetARock/GetARock/Screen/AddGathering/LocationSearchResultViewController.swift
@@ -41,15 +41,14 @@ class LocationSearchResultViewController: UIViewController {
     }
     
     deinit {
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
+        removeObserversForKeyboardShow()
     }
 
     // MARK: - Method
 
     private func attribute() {
         setupTableView()
-        getKeyboardNotification()
+        addObeserversForKeyboardShow()
     }
 
     private func setupTableView() {
@@ -100,51 +99,5 @@ extension LocationSearchResultViewController: UITableViewDataSource {
         cell.addressLabel?.text = suggestedPlaces[(indexPath as NSIndexPath).row].subtitle
 
         return cell
-    }
-}
-
-// MARK: - KeyboardControl
-
-extension LocationSearchResultViewController {
-    private func getKeyboardNotification() {
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(keyboardWillShow(_:)),
-            name: UIResponder.keyboardWillShowNotification,
-            object: nil
-        )
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(keyboardWillHide(_:)),
-            name: UIResponder.keyboardWillHideNotification,
-            object: nil
-        )
-    }
-
-    @objc func keyboardWillShow(_ sender: Notification) {
-        guard let userInfo: NSDictionary = sender.userInfo as NSDictionary?,
-              let keyboardFrame: NSValue = userInfo.value(forKey: UIResponder.keyboardFrameEndUserInfoKey) as? NSValue else {
-                  return
-              }
-        let keyboardRectangle = keyboardFrame.cgRectValue
-        let keyboardHeight = keyboardRectangle.height
-
-        let contentInset = UIEdgeInsets(
-            top: 0.0,
-            left: 0.0,
-            bottom: keyboardHeight,
-            right: 0.0)
-        tableView.contentInset = contentInset
-        tableView.scrollIndicatorInsets = contentInset
-    }
-
-    @objc func keyboardWillHide(_ sender: Notification) {
-        let contentInset = UIEdgeInsets(
-                top: 0.0,
-                left: 0.0,
-                bottom: 0.0,
-                right: 0.0)
-            tableView.contentInset = contentInset
-            tableView.scrollIndicatorInsets = contentInset
     }
 }
